% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acled_auth.R
\name{acled_auth}
\alias{acled_auth}
\title{Authenticate a request using ACLED credentials via oauth client. Used within the primary acled_api() function.}
\usage{
acled_auth(req, username, password = NULL)
}
\arguments{
\item{req}{An ACLED API request.}

\item{username}{ACLED username.}

\item{password}{User password. If NULL, opens interactive prompt}
}
\value{
Returns a success message if ACLED credentials are successfully authorized
}
\description{
Pings the ACLED API token endpoint for authentication using username and password.
}
\examples{
\dontrun{
req_auth(req, username = "your_email")
}
}
\seealso{
ACLED API Access guide \url{https://acleddata.com/download/35300/}

Other API and Access: 
\code{\link{acled_access}()},
\code{\link{acled_api}()},
\code{\link{acled_deletions_api}()},
\code{\link{acled_update}()}
}
\concept{API and Access}
