% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acled_codebook}
\alias{acled_codebook}
\title{ACLED Codebook}
\format{
A data frame:
\describe{
\item{Variable}{Variable names}
\item{Description}{Text description of each variable}
\item{Values}{Text description of values for each variable}}
}
\usage{
acled_codebook
}
\description{
Codebook for ACLED data
}
\keyword{datasets}
