% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acled_countries}
\alias{acled_countries}
\title{ACLED Countries}
\format{
A data frame:
\describe{
\item{country}{Country names}
\item{region_name}{Region names}
\item{start_year}{First year coded by ACLED}}
}
\usage{
acled_countries
}
\description{
ACLED country names, regions, and coding start year
}
\keyword{datasets}
