% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acled_coverage}
\alias{acled_coverage}
\title{ACLED country coverage}
\format{An object of class \code{data.frame} with 93 rows and 3 columns.}
\source{
\url{https://www.acleddata.com/wp-content/uploads/dlm_uploads/2018/02/Country-and-Time-Period-coverage_updatedJune2019-1.pdf}
}
\usage{
acled_coverage
}
\description{
A dataset containing the start of ACLED data coverage for a country
}
\keyword{datasets}
