% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{acled_data_path}
\alias{acled_data_path}
\title{Data path}
\usage{
acled_data_path(...)
}
\arguments{
\item{...}{additional elements to splice into path, e.g. file name}
}
\description{
Path to folder containing data
}
