% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acled_event_categories}
\alias{acled_event_categories}
\title{ACLED Event Categories}
\format{
A data frame:
\describe{
\item{event_type}{ACLED event type}
\item{sub_event_type}{ACLED sub-event type}
\item{political_violence}{Dummy indicator for whether sub-event type falls within political violence}
\item{organized_political_violence}{Dummy indicator for whether sub-event type falls within organized political violence}
\item{disorder}{Dummy indicator for whether sub-event type falls within disorder}
\item{demonstrations}{Dummy indicator for whether sub-event type falls within demonstrations}}
}
\usage{
acled_event_categories
}
\description{
ACLED event and sub-event types, grouped by category
}
\seealso{
Other Data: 
\code{\link{acled_codebook}},
\code{\link{acled_countries}},
\code{\link{acled_interaction_codes}},
\code{\link{acled_multipliers}},
\code{\link{acled_old_deletion_dummy}},
\code{\link{acled_old_dummy}},
\code{\link{acled_regions}}
}
\concept{Data}
\keyword{datasets}
