% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acled_help.R
\name{acled_help}
\alias{acled_help}
\title{ACLED's dataset help}
\usage{
acled_help(column = NULL)
}
\arguments{
\item{column}{string. A string with the name of the column for which to consult methodology.}
}
\value{
An acledR vignette
}
\description{
The function allows users to consult methodological vignettes for the different columns in ACLED's dataset
}
\examples{
\dontrun{

## Get information about the fatalities column
acled_help("fatalities")
}
}
\seealso{
\itemize{
\item ACLED resources and methodology guides \url{https://acleddata.com/resources/}
}

Other Helpers: 
\code{\link{acled_rounding}()}
}
\concept{Helpers}
