% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acled_interaction_codes}
\alias{acled_interaction_codes}
\title{ACLED interaction codes}
\format{
A data frame:
\describe{
\item{Inter1/Inter2}{Actor type}
\item{Numeric Code}{Numeric equivalent found in the inter1 and inter2 column. }}
}
\usage{
acled_interaction_codes
}
\description{
ACLED interaction and actor types
}
\seealso{
Other Data: 
\code{\link{acled_codebook}},
\code{\link{acled_countries}},
\code{\link{acled_event_categories}},
\code{\link{acled_multipliers}},
\code{\link{acled_old_deletion_dummy}},
\code{\link{acled_old_dummy}},
\code{\link{acled_regions}}
}
\concept{Data}
\keyword{datasets}
