% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acled_multipliers}
\alias{acled_multipliers}
\title{ACLED Multipliers}
\format{
A data frame:
\describe{
\item{country}{Country names}
\item{bin}{Bin of event frequency}
\item{year}{Year corresponding to the bin}
\item{avg_month_bin}{Average monthly of the bin}
}
}
\usage{
acled_multipliers
}
\description{
A dataframe with additional information for each country, only for the purpose of estimating events.
}
\seealso{
Other Data: 
\code{\link{acled_codebook}},
\code{\link{acled_countries}},
\code{\link{acled_event_categories}},
\code{\link{acled_interaction_codes}},
\code{\link{acled_old_deletion_dummy}},
\code{\link{acled_old_dummy}},
\code{\link{acled_regions}}
}
\concept{Data}
\keyword{datasets}
