% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acled_regen.R
\name{acled_regen}
\alias{acled_regen}
\title{Regenerating ACLED's deleted columns}
\usage{
acled_regen(df, column)
}
\arguments{
\item{df}{data frame. ACLED's data frame where the regenerated column will be created}

\item{column}{character string. Name of the column to regenerate. See details for info on available columns}
}
\value{
Tibble with the regenerated column addedd at the end.
}
\description{
This formula allows to regenerate columns (when applicable) that may have been deleted from the dataset.
}
\note{
Please note that as of today, the one columns available to be regenerated are
\itemize{
\item event_id_no_cnty
\item iso3
}
}
