% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acled_regions}
\alias{acled_regions}
\title{ACLED Regions}
\format{
A data frame:
\describe{
\item{region}{Region number}
\item{region_name}{Region names}
\item{first_event_date}{First date (yyyy-mm-dd) coded by ACLED}}
}
\usage{
acled_regions
}
\description{
ACLED Region Names, Region Numbers, and Coding Start Dates
}
\keyword{datasets}
