% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acled_report_api.R
\name{acled_report_api}
\alias{acled_report_api}
\alias{acled_country_report_api}
\title{Automatic country report with API}
\usage{
acled_country_report_api(
  country_report = NULL,
  regional_data = TRUE,
  region = NULL,
  start_date_report = "1997-01-01",
  end_date_report = Sys.Date(),
  output_style = "HTML",
  email = NULL,
  key = NULL,
  acled_access = TRUE,
  prompt = TRUE
)
}
\arguments{
\item{country_report}{character string. A country from which to base the report.}

\item{regional_data}{boolean. True or False option to include (when TRUE) regional comparision in the report. Requires larger API calls.}

\item{region}{character string. A region which from which to base the report. It still requires a country.}

\item{start_date_report}{character string. The initial date (YYYY-MM-DD) from which to generate the report.}

\item{end_date_report}{character string. Last date (YYYY-MM-DD) from which to finish the report.}

\item{output_style}{character string. Desired format for the markdown file, recommended HTML or PDF. Defaults to HTML.}

\item{email}{character string. Email associated with your ACLED account registered at \url{https://developer.acleddata.com}.}

\item{key}{character string. Access key associated with your ACLED account registered at \url{https://developer.acleddata.com}.}

\item{acled_access}{logical. If TRUE (default), you have used the acled_access function and the email and key arguments are not required.}

\item{prompt}{logical. If TRUE (default), users will receive an interactive prompt providing information about their call (countries requested, number of country-days, and number of API calls required) and asking if they want to proceed with the call. If FALSE, the call continues without warning, but the call is split and returns a message specifying how many calls are being made.}
}
\value{
Returns a markdown file stored in the project's current working directory.
}
\description{
Generate automatic exploratory reports of ACLED data by utilizing data directly from ACLED's API
}
\examples{
\dontrun{

## Generate report regarding Argentina's events throughout August-December 2022
acled_country_report_api(country_report = "Argentina", start_date_report = "2022-07-31",end_date_report = "2023-01-01",
                         output_style = "HTML", prompt=FALSE)
}
}
