% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acled_sample}
\alias{acled_sample}
\title{ACLED event data sample}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 30 columns.}
\usage{
acled_sample
}
\description{
A sample of the ACLED event data
}
\keyword{datasets}
