% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acled_transform.R
\name{acled_transform}
\alias{acled_transform}
\title{Transform ACLED data from wide to long}
\usage{
acled_transform(data, type = "full_actors")
}
\arguments{
\item{data, }{Dataframe or tibble containing your dataset.}

\item{type, }{character string. One of five types: full_actors, main_actors, assoc_actors, source, or all.
\itemize{
\item full_actors: All actor and associated actor columns
\item main_actors: Actor 1 and Actor 2 columns
\item assoc_actors: All associated actor columns
\item source: The source column becomes monadic
\item all: All actor, associated actor and source columns become monadic
}}
}
\value{
A tibble with the data transformed into long form.
}
\description{
Function to convert your ACLED's API calls (if dyadic) into desired monadic forms.
}
\examples{
\dontrun{
#argen_acled <- acled_api(countries = "Argentina",start_date = "2022-01-01",
#                          end_date="2022-02-01", acled_access = T, prompt = F)

#argen_acled_long_actors <- acled_transform(argen_acled,
#                                            type = "full_actor") # Transforming the data

#nrow(argen_acled_long_actors) # Number of rows in the dataset
#[1] 263 # Long form

#nrow(argen_acled) ) # Number of rows in the dataset
#[1] 145 # Wide form
}
}
\seealso{
Other Data Manipulation: 
\code{\link{acled_regen}()},
\code{\link{filter_actors}()},
\code{\link{filter_event_types}()}
}
\concept{Data Manipulation}
