% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acled_update.R
\name{acled_update}
\alias{acled_update}
\title{Updating your ACLED dataset}
\usage{
acled_update(
  df,
  start_date = min(df$event_date),
  end_date = max(df$event_date),
  countries = NULL,
  regions = NULL,
  event_types = NULL,
  acled_access = TRUE,
  email = NULL,
  key = NULL,
  deleted = TRUE,
  prompts = T
)
}
\arguments{
\item{countries}{string. The countries by which to filter.}

\item{regions}{string. The regions by which to filter.}

\item{event_types}{string. The event types by which to filter.}

\item{acled_access}{logical. If TRUE (default), you have used the acled_access function and the email and key arguments are not required.}

\item{email}{character string. Email associated with your ACLED account registered at \url{https://developer.acleddata.com}.}

\item{key}{character string. Access key associated with your ACLED account registered at \url{https://developer.acleddata.com}.}

\item{deleted}{logical. If TRUE (default), the function will also remove deleted events using acled_deletions_api().}

\item{prompt}{logical. If TRUE (default), users will receive an interactive prompt providing information about their call (countries requested, number of country-days, and number of API calls required) and asking if they want to proceed with the call. If FALSE, the call continues without warning, but the call is split and returns a message specifying how many calls are being made.}
}
\description{
This function is meant to help you keep your dataset updated, by automatically checking for new and modified events, as well as deleted events (if deleted = TRUE).
Note: The function makes new api calls to gather new and modified events. See its vignettes in: \url{https://acled.github.io/acledR/articles/acled_deletions_api.html}
}
\seealso{
Other API and Access: 
\code{\link{acled_access}()},
\code{\link{acled_api}()},
\code{\link{acled_deletions_api}()}
}
\concept{API and Access}
