% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acledR-package.R
\docType{package}
\name{acledR-package}
\alias{acledR}
\alias{acledR-package}
\title{acledR: Manipulate 'ACLED' Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The package allows users to easily interact with 'ACLED' data by providing wrappers for the API and other functions to manipulate 'ACLED' data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ACLED/acledR}
  \item \url{https://acled.github.io/acledR/}
  \item \url{https://dtacled.github.io/acledR/}
  \item Report bugs at \url{https://github.com/ACLED/acledR/issues}
}

}
\author{
\strong{Maintainer}:  ACLED \email{data@acleddata.com} [copyright holder]

}
\keyword{internal}
