% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curated-data-urls.R
\name{curated_data_urls}
\alias{curated_data_urls}
\title{Curated data URLs}
\usage{
curated_data_urls(which = NULL)
}
\arguments{
\item{which}{Which region's URL to return; NULL (default) will return all}
}
\value{
a named list containing the URL(s)
}
\description{
URLs for the ACLED curated data files
}
