% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_acled}
\alias{download_acled}
\title{Download ACLED data}
\usage{
download_acled(to_dir = NULL, which = "all")
}
\arguments{
\item{to_dir}{destination directory, if NULL set options}

\item{which}{which region to download, by default all.}
}
\description{
Download the curated ACLED .xlsx data files
}
