% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_actors.R
\name{filter_actors}
\alias{filter_actors}
\title{Filter ACLED Data by Actor}
\usage{
filter_actors(df, actors, filter_cols = "all")
}
\arguments{
\item{df}{dataframe. ACLED data to filter.}

\item{actors}{character vector or vector of character strings. Actor name or list of actor names to use for filtering.}

\item{filter_cols}{character string. Columns to use for filtering. Default is "all", which filters based on all actor columns (i.e. actor1, actor2, assoc_actor_1, assoc_actor_2). "Primary" filters based on actor1 and actor2 column only.}
}
\value{
Returns a tibble of ACLED data filtered to selected actors.
}
\description{
This function allows users to filter ACLED data by actor.
}
\seealso{
Other Data Manipulation: 
\code{\link{calculate_actor_index}()},
\code{\link{filter_event_types}()},
\code{\link{generate_counts}()},
\code{\link{generate_movers}()}
}
\concept{Data Manipulation}
