% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_event_types.R
\name{filter_event_types}
\alias{filter_event_types}
\title{Filter ACLED data by event type categories}
\usage{
filter_event_types(data, acled_categories)
}
\arguments{
\item{data}{ACLED data}

\item{acled_categories}{character vector of ACLED event type categories. Options include: political_violence, organized_political_violence, disorder, and demonstrations. More than one category may be requested at one time (e.g., acled_categories = c("organized_political_violence", "disorder")).}
}
\value{
Returns a tibble including only ACLED events that fall within the requested event categories, including new columns (dummy variables) indicating which categories each event falls under.
}
\description{
Filter ACLED data by event type categories
}
\examples{
\dontrun{

# Request data for 3 countries from the ACLED API
df_events <- acled_api(countries = c("Nigeria", "Benin", "Senegal"),
                       start_date = "2022-01-01",
                       end_date = "2022-09-30",
                       monadic = F,
                       acled_access = TRUE)

# Filter to only ACLED demonstrations
df_demonstrations <- filter_event_types(data = df_events, acled_categories = c("demonstrations"))

# This returns a tibble with only peaceful protests, protests with intervention, violence demonstrations, and excessive force against protesters.
df_demonstrations

# Users may also filter by multiple event categories
df_demonstrations_disorder <- filter_event_types(data = df_events, acled_categories = c("demonstrations", "disorder"))

# Which returns all rows that are either demonstrations and/or disorder events
df_demonstrations_disorder

}
}
\seealso{
Other Data Manipulation: 
\code{\link{calculate_actor_index}()},
\code{\link{generate_counts}()},
\code{\link{generate_movers}()}
}
\concept{Data Manipulation}
