% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_counts.R
\name{generate_counts}
\alias{generate_counts}
\title{Generate event counts from ACLED data}
\usage{
generate_counts(
  data,
  event_type = NULL,
  unit_id,
  time_id,
  time_target,
  start_date = NULL,
  end_date = NULL,
  add_unit_ids = NULL
)
}
\arguments{
\item{data}{ACLED data}

\item{event_type}{Event types to include}

\item{unit_id}{Unit variable}

\item{time_id}{Temporal variable}

\item{time_target}{Target temporal unit}

\item{start_date}{Earliest date to include}

\item{end_date}{Latest date to include}

\item{add_unit_ids}{Option to add in units with no events at certain time periods}
}
\value{
Returns a tibble grouped by unit_id
}
\description{
Generate event counts from ACLED data
}
