% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepInTable.R
\name{grepInTable}
\alias{grepInTable}
\title{grep for each element of a character vector patter in referenceTable}
\usage{
grepInTable(
  pattern,
  referenceTable = rworldmap::countrySynonyms[, -1],
  ignore.case = TRUE
)
}
\arguments{
\item{pattern}{A character vector containing regular expressions to search for matches in refereanceTable[, -1] and return the corresponding element of referenceTable[, 1]}

\item{referenceTable}{An object to search for pattern in as.data.frame(referenceTable)[-1]}

\item{ignore.case}{FALSE to search for pattern; TRUE to search for toupper(pattern)}
}
\value{
A character vector containing matches or error messages with names = pattern
}
\description{
grep for each element of a character vector patter in referenceTable
}
\examples{
AfgNethAnt <- grepInTable(c('Afghanistan', 'Netherlands Antilles'))
AfgNthAnt <- c(Afghanistan="AFG", "Netherlands Antilles" = 
       "Found in row 179 of referenceTable with column 1 = ''") 
all.equal(AfgNethAnt, AfgHtnAnt)
}
