% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_acled}
\alias{read_acled}
\title{Read ACLED data}
\usage{
read_acled(data_path = NULL)
}
\arguments{
\item{data_path}{path to data folder, if NULL (default) set by consulting
[acled_data_path()].}
}
\description{
Read ACLED data into memory
}
