% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acled_access.R
\name{acled_access}
\alias{acled_access}
\title{Store your ACLED access information into your session.}
\usage{
acled_access(email, key)
}
\arguments{
\item{email}{This is the email which you register in the ACLED Access portal}

\item{key}{This is the key generated by the ACLED Access portal.}
}
\description{
Simple function to store (through \code{Sys.setenv()} ) ACLED access keys for the acled_api() function.
}
\examples{
\dontrun{
acled_access("joe.doe\\\\@email.com","random_sequence_of_numbers_and_letters")
Sys.getenv("acled_key")
[1] "random_sequence_of_numbers_and_letters"
}
}
\seealso{
ACLED API Access guide. \url{https://acleddata.com/acleddatanew//wp-content/uploads/dlm_uploads/2021/11/API-User-Guide_Feb2022.pdf}

Other API and Access: 
\code{\link{acled_api}()}
}
\concept{API and Access}
