% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acled_api.R
\name{acled_api}
\alias{acled_api}
\title{Request data from ACLED API}
\usage{
acled_api(
  email,
  key,
  countries = NULL,
  regions = NULL,
  start_date = NULL,
  end_date = NULL,
  monadic = FALSE
)
}
\arguments{
\item{email}{character string. Email associated with your ACLED account registered at https://developer.acleddata.com.}

\item{key}{character string. Access key associated with your ACLED account registered at https://developer.acleddata.com.}

\item{countries}{character vector. Default is NULL, which will return events for all countries. Pass a vector of country names to retrieve events from specific countries. The list of ACLED country names may be found via acledR::acled_countries,}

\item{regions}{vector of region names (character) or region codes (numeric). Default is NULL, which will return events for all regions.  Pass a vector of regions names or codes to retrieve events from countries within specific regions. The list of ACLED regions may be found via acledR::acled_regions,}

\item{start_date}{character string. Format 'yyyy-mm-dd'. The earliest date for which to return events. The default is NULL, which will return events from all available time periods. If 'start_date' is NULL, 'end_date' must also be NULL.}

\item{end_date}{character string. Format 'yyyy-mm-dd'. The latest date for which to return events. The default is NULL, which will return events from all available time periods. If 'end_date' is NULL, 'start_date' must also be NULL.}

\item{monadic}{logical. If FALSE (default), returns dyadic data. If TRUE, returns monadic data.}
}
\value{
Returns a tibble of of ACLED events.
}
\description{
Processes requests to ACLED's API
}
