% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acled_countries}
\alias{acled_countries}
\title{ACLED Countries}
\format{
A data frame:
\describe{
\item{country}{Country names}
\item{region}{Region names}
\item{start_year}{First year coded by ACLED}}
}
\usage{
acled_countries
}
\description{
ACLED country names, regions, and coding start year
}
\seealso{
Other Data: 
\code{\link{acled_codebook}},
\code{\link{acled_event_categories}},
\code{\link{acled_interaction_codes}},
\code{\link{acled_multipliers}},
\code{\link{acled_old_deletion_dummy}},
\code{\link{acled_old_dummy}},
\code{\link{acled_regions}}
}
\concept{Data}
\keyword{datasets}
