% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acledR-package.R
\docType{package}
\name{acledR-package}
\alias{acledR}
\alias{acledR-package}
\title{acledR: Manipulate ACLED Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The official ACLED R package for extracting and manipulating ACLED data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dtacled.github.io/acledR/}
  \item Report bugs at \url{https://github.com/dtacled/acledR/issues}
}

}
\author{
\strong{Maintainer}: Armed Conflict Location and Event Data ACLED \email{data@acleddata.com} [copyright holder]

Authors:
\itemize{
  \item Trey Billing \email{t.billing@acleddata.com}
  \item Lucas Fagliano
}

Other contributors:
\itemize{
  \item Katayoun Kishi [contributor]
}

}
\keyword{internal}
