% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acledR-package.R
\docType{package}
\name{acledR-package}
\alias{acledR}
\alias{acledR-package}
\title{acledR: Manipulate ACLED Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools working with data from ACLED (Armed Conflict Location and Event Data). Functions include simplified access to ACLED's API (\url{https://apidocs.acleddata.com/}), methods for keeping local versions of ACLED data up-to-date, and functions for common ACLED data transformations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dtacled.github.io/acledR/}
  \item Report bugs at \url{https://github.com/dtacled/acledR/issues}
}

}
\author{
\strong{Maintainer}: Trey Billing \email{t.billing@acleddata.com}

Authors:
\itemize{
  \item Lucas Fagliano \email{lucasfagliano@gmail.com}
}

Other contributors:
\itemize{
  \item Armed Conflict Location and Event Data ACLED \email{data@acleddata.com} [copyright holder]
  \item Katayoun Kishi [contributor]
}

}
\keyword{internal}
