% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_actor_index.R
\name{calculate_actor_index}
\alias{calculate_actor_index}
\title{Calculate actor concentration indices from ACLED data}
\usage{
calculate_actor_index(events, method = "Effective actors")
}
\arguments{
\item{events}{Vector of outcomes per actor (i.e., event counts or fatalities).}

\item{method}{Index method. Default is "Effective actors", which is an Inverse Simpson Index. "Concentration" calculates a Herfindahl–Hirschman Index.}
}
\value{
Returns a data.frame of the index value, number of unique actors, and average number of events per actor.
}
\description{
Calculate actor concentration indices from ACLED data
}
