% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_actors.R
\name{filter_actors}
\alias{filter_actors}
\title{Filter ACLED Data by Actor}
\usage{
filter_actors(df, actors, filter_cols = "all")
}
\arguments{
\item{df}{dataframe. ACLED data to filter.}

\item{actors}{character vector or vector of character strings. Actor name or list of actor names to use for filtering.}

\item{filter_cols}{character string. Columns to use for filtering. Default is "all", which filters based on all actor columns (i.e. actor1, actor2, assoc_actor_1, assoc_actor_2). "Primary" filters based on actor1 and actor2 column only.}
}
\value{
Returns a tibble of ACLED data filtered to selected actors.
}
\description{
This function allows users to filter ACLED data by actor.
}
\examples{
\dontrun{

# Request events from Yemen
df_events <- acled_api(countries = "Yemen",
                       start_date = "2022-01-01",
                       end_date = "2022-07-30",
                       monadic = F,
                       acled_access = TRUE)

# Create vector of actors of interest
yemen_actors <- c('Military Forces of Yemen (2016-) Supreme Political Council',
                  'Military Forces of Yemen (2012-)')

# Filter to selected actors
filtered_df <- filter_actors(df = df_events, actors = yemen_actors, filter_cols = 'all')
}
}
\seealso{
Other Data Manipulation: 
\code{\link{acled_filter_event_types}()},
\code{\link{acled_regen}()},
\code{\link{acled_transform}()}
}
\concept{Data Manipulation}
