% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_movers.R
\name{generate_movers}
\alias{generate_movers}
\title{Generate moving statistics}
\usage{
generate_movers(
  data,
  var,
  unit_id,
  time_id,
  slide_funs,
  slide_periods,
  na.rm = T,
  complete = T
)
}
\arguments{
\item{data}{ACLED data.}

\item{unit_id}{Unit variable.}

\item{time_id}{Temporal variable.}

\item{slide_funs}{Requested moving statistics. Character vector with options including mean, median, min, and max.}

\item{slide_periods}{How many periods in the past to summarize over. Vector of one or more integers. Inf includes all previous periods.}

\item{na.rm}{Whether to include NAs in the calculations.}
}
\value{
Returns a tibble grouped by unit_id.
}
\description{
Generate moving statistics
}
